reduce warnings

+++ src/helper/HPrint.cpp
@@ -6,15 +6,16 @@
 #include "HPrint.h"
 
 
-
 void Print(PrintType printType, std::string const& message) {
 #ifndef _DEBUG
-	if (printType == PrintType::ONLY_DEBUG) { return; }
+    if (printType == PrintType::ONLY_DEBUG) {
+        return;
+    }
 #endif // _DEBUG
 
-	std::string const typeS   { GetPrintTypeString(printType) };
-	std::string const toExport{ typeS + " " + message + '\n'  };
+    std::string const typeS{ GetPrintTypeString(printType) };
+    std::string const toExport{ typeS + " " + message + '\n' };
 
-	TryExport(toExport, printType);
-	std::cout << std::setw(GetPrintTypeString(longestType).size()) << typeS << ' ' << message << '\n';
+    TryExport(toExport, printType);
+    std::cout << std::setw(static_cast<int>(GetPrintTypeString(longestType).size())) << typeS << ' ' << message << '\n';
 }
